refactor(Lotto): Lotto 클래스를 model 패키지로 이동
