+++ src/main/kotlin/model/Lotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package model
 
 class Lotto(private val numbers: List<Int>) {
     init {
@@ -6,4 +6,7 @@ class Lotto(private val numbers: List<Int>) {
     }
 
     // TODO: 추가 기능 구현
+    override fun toString(): String {
+        return numbers.toString()
+    }
 }
