+++ src/main/kotlin/model/LottoMachine.kt
@@ -1,7 +1,6 @@
 package model
 
 import camp.nextstep.edu.missionutils.Randoms
-import lotto.Lotto
 
 object LottoMachine {
